added apple class (has a buggy random location),  made classes draw their own objects (snakebody wont draw), removed board.cpp

+++ SFML-proj/Location.h
@@ -23,7 +23,5 @@ struct Location
 		return { -x,-y };
 	}
 
-	int x, y;
-
-
+	float x, y;
 };
